+++ terminal.h
@@ -554,6 +554,7 @@ struct terminal {
         struct buffer *last_buf;     /* Buffer we rendered to last time */
         bool was_flashing;           /* Flash was active last time we rendered */
         bool was_searching;
+        bool was_focused;
 
         size_t search_glyph_offset;